Trying to export in external class
